Create Contains()
Returns true or false depending if a certain value is in the linked
list. This requires recursion, and two base cases; if the current node's
value is equal to the specified value, return true, and after that, if
the end of the list is reached (the current node doesn't have a next
node), return false.