Create InsertAt()
Recursively loop through just before the node at the specified index.
Then create a new node, and insert in the linked list by assigning
the new node's next node to the following node in the linkedlist (which
is node.getNextNodeLink().getNextNodeLink()), and assign the new node to
be the next node of the previous node.